+++ ngbatis-demo/src/main/java/ye/weicheng/ngbatis/demo/repository/TestGeoDao.java
@@ -1,0 +1,17 @@
+package ye.weicheng.ngbatis.demo.repository;
+
+// Copyright (c) 2025 All project authors. All rights reserved.
+//
+// This source code is licensed under Apache 2.0 License.
+
+import org.nebula.contrib.ngbatis.proxy.NebulaDaoBasic;
+import ye.weicheng.ngbatis.demo.pojo.TestGeo;
+
+/**
+ * @author yeweicheng
+ * @since 2025-05-17 11:41
+ * <br>Now is history!
+ */
+public interface TestGeoDao extends NebulaDaoBasic<TestGeo, String> {
+
+}